+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All Notable changes to `Password` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## 0.2.0 - 2017-02-27
+
+### Added
+- Info class
+
+### Deprecated
+- Nothing
+
+### Fixed
+- Nothing
+
+### Removed
+- `getCost` from Password class
+- `getAlgorithm` from Password class
+
+### Security
+- Nothing
+
 ## 0.1.0 - 2017-02-23
 
 ### Added